% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_vec.R
\name{collapse_vec}
\alias{collapse_vec}
\title{Title  Collapse a vector into a string, separated by a pipe character (|).}
\usage{
collapse_vec(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A string.
}
\description{
Description  Collapse a vector into a string, separated by a pipe character (|).
}
\examples{
x <- c("a", "b", "c")
collapse_vec(x)
}
