% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_conf_intervals.R
\name{combine_conf_intervals}
\alias{combine_conf_intervals}
\title{Create a string with CI given estimate and confidence interval
Combines a point estimate, lower bound, and upper bound into a string in the format}
\usage{
combine_conf_intervals(estimate, lower, upper)
}
\arguments{
\item{estimate}{The point estimate.}

\item{lower}{The lower bound of the confidence interval.}

\item{upper}{The upper bound of the confidence interval.}
}
\value{
A string in this format  \verb{estimate [lower upper]}.
}
\description{
Create a string with CI given estimate and confidence interval
Combines a point estimate, lower bound, and upper bound into a string in the format
}
\examples{

combine_conf_intervals(0.5, 0.4, 0.6)

}
