% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_save_workbook.R
\name{create_save_workbook}
\alias{create_save_workbook}
\title{Title}
\usage{
create_save_workbook(list_of_dfs, path_name, sheet_names)
}
\arguments{
\item{list_of_dfs}{A list of data frames to write to the workbook.}

\item{path_name}{The path to the Excel workbook to create.}

\item{sheet_names}{A character vector of names for the sheets in the workbook.}
}
\value{
The workbook object.
}
\description{
Description
}
\examples{
df1 <- data.frame(x = 1:10, y = 11:20)
df2 <- data.frame(a = letters[1:5], b = letters[6:10])
list_of_dfs <- list(df1, df2)

# Specify the names of the sheets
sheet_names <- c("Sheet1", "Sheet2")

#create temporary folder 
path_name <- tempfile(pattern = "create_work_book") |>
  normalizePath(winslash = "/")

dir.create(path_name)

# create file name
file_name <- file.path(path_name, "test.xlsx")

# Create the workbook
create_save_workbook(list_of_dfs,
                     path_name = file_name, 
                     sheet_names)
file.remove(file_name)
unlink(path_name)
}
