% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distances between two geo-coordinates}
\usage{
distance(lat1, lon1, lat2, lon2, unit)
}
\arguments{
\item{lat1}{Latitude and Longitude of point 1 (in decimal degrees)}

\item{lon1}{Latitude and Longitude of point 1 (in decimal degrees)}

\item{lat2}{Latitude and Longitude of point 2 (in decimal degrees)}

\item{lon2}{Latitude and Longitude of point 2 (in decimal degrees)}

\item{unit}{'M' for miles (default), 'K' for kilometers, 'N' for nautical miles}
}
\value{
Distance betwen two points
}
\description{
This function is just a copy of javascript version available on
https://www.geodatasource.com/developers/javascript.
(under LGPLv3 license).
}
\details{
Official Web site: http://www.geodatasource.com
}

