% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_tidy_model.R
\name{DT_tidy_model}
\alias{DT_tidy_model}
\title{Title  Produce a nice data table from tidy model object
Description This function takes a tidy model object and converts it into a formatted data table. It can use custome function eg knitr kable}
\usage{
DT_tidy_model(model, round_digts = 4, output_function = "data_table")
}
\arguments{
\item{model}{The tidy model object.}

\item{round_digts}{An integer specifying the number of digits to round the numeric columns. Default is 4.}

\item{output_function}{A string specifying the output function to use. Default is "data_table".}
}
\value{
An object of class DT, representing the formatted data table.
}
\description{
Title  Produce a nice data table from tidy model object
Description This function takes a tidy model object and converts it into a formatted data table. It can use custome function eg knitr kable
}
\examples{
## model from glm examples
library(MASS)
data(anorexia)

anorex.1 <- glm(Postwt ~ Prewt + Treat + offset(Prewt),
                family = gaussian, data = anorexia)


DT_tidy_model(anorex.1)
}
