% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\docType{data}
\name{locations}
\alias{locations}
\title{Lon/Lat Coordinates of Zip-Codes/Cities}
\usage{
data(locations)
}
\description{
Converted from http://download.geonames.org/export/zip/allCountries.zip
See http://www.geonames.org/ for more.
}

