% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations2.R
\docType{data}
\name{locations2}
\alias{locations2}
\title{Lon/Lat Coordinates for Zip-Codes}
\usage{
data(locations2)
}
\description{
Converted from http://download.geonames.org/export/zip/allCountries.zip
See http://www.geonames.org/ for more.
}

