\name{mtools-package}
\alias{mtools-package}
\alias{mtools}
\docType{package}
\title{Tools for migrate-n output analysis}
\description{Tools for migrate-n model selection from output files.}
\details{
\tabular{ll}{
Package: \tab mtools\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 10-02-2017\cr
License: \tab GPL-2 | GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{Lizandro N. Peraza-Flores

Maintainer: Lizandro N. Peraza-Flores email= lizandropf@gmail.com
}
\references{
For details on methods for summary statistics selection, see:

Beerli, P. & Felsenstein, J. (2001) Maximum likelihood estimation of a
migration matrix and effective population sizes in n subpopulations by
using a coalescent approach. \emph{Proceedings of the National Academy
of Sciences}, 98, 4563--4568.\cr\cr
Beerli, P. & Palczewski, M. (2010) Unified framework to evaluate panmixia
and migration direction among multiple sampling locations.
\emph{Genetics}, \bold{185}, 313--326.\cr\cr

Details of migrate-n analysis appear in:

Beerli, P. & Felsenstein, J. (2001) Maximum likelihood estimation of a
migration matrix and effective population sizes in n subpopulations by
using a coalescent approach. \emph{Proceedings of the National Academy
of Sciences}, 98, 4563--4568.\cr\cr
}
\keyword{
  package
  methods
}
