% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvis.R
\name{mvis}
\alias{mvis}
\title{Title}
\usage{
mvis(
  df,
  x,
  y = NULL,
  x_label = "NULL",
  y_label = "NULL",
  colorset = 1,
  categoryCol = "NULL"
)
}
\arguments{
\item{df}{A dataframe from which to use variables from.}

\item{x}{variable from the dataframe to create a visualization. A univariate is only x is provided}

\item{y}{variable from the dataframe to create a visualization.}

\item{x_label}{Label to be displayed representing the x variable.}

\item{y_label}{Label to be displayed representing the y variable.}

\item{colorset}{Numerical value for selecting a color set '1, 2, 3, 4'}

\item{categoryCol}{Discrete variable for setting plot colors based on it}
}
\value{
Returns a visualizations that suit variables given.
}
\description{
Title
}
\examples{
Example: df, df$height, df$weight, 'Height', 'Weight', 4, df$gender

}
