% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_content.R
\name{plotProfile_content}
\alias{plotProfile_content}
\title{Plot content consumed with barplots}
\usage{
plotProfile_content(data, uid)
}
\arguments{
\item{data}{A list of data.frames with the 'media' format}

\item{uid}{an interesting id to be marked}
}
\value{
gg object
}
\description{
This function ggplots number of consumed hours of content per content title
}

