% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_localisation.R
\name{plotProfile_place}
\alias{plotProfile_place}
\title{Plot zipcodes}
\usage{
plotProfile_place(data, uid)
}
\arguments{
\item{data}{A list of data.frames with the 'media' format}

\item{uid}{an interesting id to be marked}
}
\value{
gg object
}
\description{
This function leafplots zipcodes
}

