% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_week.R
\name{plotProfile_week}
\alias{plotProfile_week}
\title{Plot content consumed during the week with barplots}
\usage{
plotProfile_week(data, uid)
}
\arguments{
\item{data}{A list of data.frames with the 'media' format}

\item{uid}{an interesting id to be marked}
}
\value{
gg object
}
\description{
This function ggplots number of consumed hours of content per day and hour
}

