% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZIP.R
\name{plotZIP}
\alias{plotZIP}
\title{Plot zip-codes}
\usage{
plotZIP(codes)
}
\arguments{
\item{code1}{Zip code 1}

\item{code2}{Zip code 2}

\item{unit}{'M' for miles (default), 'K' for kilometers, 'N' for nautical miles}
}
\value{
Distance betwen two zip codes
}
\description{
This function is just a copy of javascript version available on
https://www.geodatasource.com/developers/javascript.
(under LGPLv3 license).
}
\details{
Official Web site: http://www.geodatasource.com
}

