% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_all_num_cols.R
\name{round_all_num_cols}
\alias{round_all_num_cols}
\title{Title Round all numeric columns to specified digits}
\usage{
round_all_num_cols(df, round_digits = 2)
}
\arguments{
\item{df}{A data frame}

\item{round_digits}{The number of digits to round to}
}
\value{
A data frame with all numeric columns rounded to the specified number of digits
}
\description{
Description A function to round all numeric columns of a data frame
}
\examples{
round_all_num_cols(iris) \%>\%
  head()
}
