% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table.R
\name{data_table}
\alias{data_table}
\title{Title Custom function for datatables}
\usage{
data_table(df, pageLength = 50)
}
\arguments{
\item{df}{A data frame}

\item{pageLength}{An integer, number of rows displayed}
}
\value{
a an object of class DT

Object of class DT
}
\description{
Description A function to produce beautiful tables
}
\examples{
iris_head  <- head(iris)
data_table(iris_head)
}
