% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_tidy_model.R
\name{DT_tidy_model}
\alias{DT_tidy_model}
\title{Title  Produce a nice data table from tidy model object
Description This function takes a tidy model object and converts it into a formatted data table. It can use custome function eg knitr kable}
\usage{
DT_tidy_model(
  model,
  round_digts = 4,
  tidy_function = "tidy",
  output_function = "data_table",
  coefficient_name = "estimate",
  coefficient_name_new = NULL,
  exp_estimate = FALSE,
  coefficient_ci = FALSE,
  confint_level = 0.95
)
}
\arguments{
\item{model}{The tidy model object.}

\item{round_digts}{An integer specifying the number of digits to round the numeric columns. Default is 4.}

\item{tidy_function}{a function that returns tibble, data frame from a model object eg tidy, glance functions from broom}

\item{output_function}{A string specifying the output function to use. Default is "data_table".}

\item{coefficient_name}{The name of the column with coeffcients}

\item{coefficient_name_new}{In case you want a custom name for your coefficient eg ODDS Ratio 95\%CI}

\item{exp_estimate}{in case you have logistic regression and you want odds ratio}

\item{coefficient_ci}{whether to add coefficent CI}

\item{confint_level}{Confidence level to be used}
}
\value{
An object of class output_function, ie kable, DT etc representing the formatted data table.
}
\description{
Title  Produce a nice data table from tidy model object
Description This function takes a tidy model object and converts it into a formatted data table. It can use custome function eg knitr kable
}
\examples{
## model from glm examples
data("diabetes")
library(MASS)
library(knitr)
library(broom)
model <- glm(Outcome ~ .,
             family = binomial(),
             data = diabetes)


DT_tidy_model(model, round_digts = 4,
              tidy_function = "tidy", 
              output_function = "kable", 
              coefficient_name = "estimate",
              coefficient_name_new = "ODDS Ratio 95\% CI",
              exp_estimate = TRUE, 
              coefficient_ci= TRUE,
              confint_level = .95)
}
