% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mTools-package.R
\docType{package}
\name{mTools-package}
\alias{mTools}
\alias{mTools-package}
\title{mTools: Functions I Normally Use For My Work And Hobies}
\description{
A package with a collection of useful functions I normally use.
}
\author{
\strong{Maintainer}: Moses Mburu \email{mmburu20@gmail.com} (\href{https://orcid.org/0000-0001-7281-583X}{ORCID})

}
\keyword{internal}
