% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{plotProfile}
\alias{plotProfile}
\title{Plot profile with vioplots}
\usage{
plotProfile(data, uid, sort.max = TRUE)
}
\arguments{
\item{data}{A data.frame in the long format with columns id, variable, value}

\item{uid}{an interesting id to be marked}

\item{sort.max}{shall variables be sorted along max value?}
}
\value{
gg object
}
\description{
This function ggplots vioplots for few variables and mark
interesting points on these distributions
}

