% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_time.R
\name{plotProfile_time}
\alias{plotProfile_time}
\title{Plot time consumed with barplots}
\usage{
plotProfile_time(data, uid)
}
\arguments{
\item{data}{A list of data.frames with the 'media' format}

\item{uid}{an interesting id to be marked}
}
\value{
gg object
}
\description{
This function ggplots number of consumed hours of content
}

